Backend/hairify/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.39.0
cryptography==42.0.5
cssselect2==0.7.0
dataclasses-json==0.6.4
Django==5.0.3
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
frozenlist==1.4.1
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-auth==2.28.2
google-cloud-aiplatform==1.43.0
google-cloud-bigquery==3.19.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.15.0
google-crc32c==1.5.0
google-generativeai==0.3.2
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpc-google-iam-v1==0.13.0
grpcio==1.62.1
grpcio-status==1.62.1
html5lib==1.1
idna==3.6
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.11
langchain-community==0.0.27
langchain-core==0.1.30
langchain-google-genai==0.0.9
langchain-google-vertexai==0.1.0
langchain-text-splitters==0.0.1
langsmith==0.1.23
lxml==5.1.1
marshmallow==3.21.1
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
orjson==3.9.15
oscrypto==1.3.0
packaging==23.2
pillow==10.2.0
proto-plus==1.23.0
protobuf==4.25.3
psycopg==3.1.18
psycopg-binary==3.1.18
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
pyHanko==0.23.2
pyhanko-certvalidator==0.26.3
PyJWT==2.8.0
pypdf==4.1.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.0.9
requests==2.31.0
rsa==4.9
shapely==2.0.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.28
sqlparse==0.4.4
svglib==1.5.1
tenacity==8.2.3
tinycss2==1.2.1
tqdm==4.66.2
types-protobuf==4.24.0.20240311
types-requests==2.31.0.20240311
typing-inspect==0.9.0
typing_extensions==4.10.0
tzlocal==5.2
uritools==4.0.2
urllib3==2.2.1
webencodings==0.5.1
xhtml2pdf==0.2.15
yarl==1.9.4
